pyproject.toml
[project]
name = "tapio"
version = "0.1.0"
description = "An assistant for Finnish immigrants"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Finntegrate Team", email = "info@finntegrate.org"}
]
dependencies = [
    "typer>=0.9.0",
    "lxml>=4.9.3",
    "html2text>=2020.1.16",
    "pyyaml>=6.0.1",
    "langchain-text-splitters>=0.3.8",
    "python-frontmatter>=1.1.0",
    "langchain-community>=0.3.21",
    "langchain>=0.3.23",
    "langchain-chroma>=0.2.3",
    "langchain-huggingface>=0.1.2",
    "gradio>=5.25.2",
    "ollama>=0.4.8",
    "pydantic>=2.11.3",
    "httpx>=0.28.1",
    "beautifulsoup4>=4.13.4",
    "google-adk>=1.7.0",
    "google-generativeai>=0.8.3",
]

[project.scripts]
tapio = "tapio.cli:app"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["tapio"]

[tool.pytest]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pyrefly>=0.17.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.5",
    "types-lxml>=2025.3.30",
    "types-pyyaml>=6.0.12.20250402",
]
